LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-button-elementFromPoint-expected.txt



FAIL CSS Overflow: elementFromPoint for ::scroll-button is scroller assert_equals: expected Element node <div id="scroller">
  <div></div>
  <div></div>
</div> but got Element node <html><head><meta charset="utf-8">
<title>CSS Overflow: e...
